# Backend/OpenAPI.yaml
openapi: 3.0.3
info:
  title: AnaECommerce Backend API
  version: v1
  description: Backend REST API for AnaECommerce with JWT authentication and role-based access.
servers:
  - url: http://localhost:5001/api
    description: Development
  - url: https://ana.runasp.net/api
    description: Production
paths:
...
  /api/users:
    get:
      summary: Get users (admin only)
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUser'
      security:
        - BearerAuth: []
      tags:
        - Users
  /api/users/{id}:
    get:
      summary: Get a user by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200': {"description": "OK", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserDto"}}}}
        '404': {"description": "Not Found"}
      security:
        - BearerAuth: []
      tags:
        - Users
    put:
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200': {"description": "OK"}
        '400': {"description": "Bad Request"}
        '404': {"description": "Not Found"}
      security:
        - BearerAuth: []
      tags:
        - Users
    delete:
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204': {"description": "No Content"}
        '404': {"description": "Not Found"}
      security:
        - BearerAuth: []
      tags:
        - Users
...
  /api/products:
    get:
      summary: Get products (anonymous)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
      responses:
        '200': {"description": "OK"}
      tags:
        - Products
  /api/products/{id}:
    get:
      summary: Get a product by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200': {"description": "OK"}
        '404': {"description": "Not Found"}
      tags:
        - Products
    post:
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '201': {"description": "Created"}
      security:
        - BearerAuth: []
      tags:
        - Products
    put:
      summary: Update product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
      responses:
        '200': {"description": "OK"}
      security:
        - BearerAuth: []
      tags:
        - Products
    delete:
      summary: Delete product
      responses:
        '204': {"description": "No Content"}
      security:
        - BearerAuth: []
      tags:
        - Products
...
  /api/categories/{id}:
    get:
      summary: Get category by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200': {"description": "OK"}
        '404': {"description": "Not Found"}
      tags:
        - Categories
    post:
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
      responses:
        '201': { "description": "Created" }
      security:
        - BearerAuth: []
      tags:
        - Categories
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
    PaginatedUser:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        totalCount:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer
    UpdateUserDto:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
...
    ProductDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        categoryId:
          type: string
        brandId:
          type: string
        stock:
          type: integer
        imageUrl:
          type: string
    CategoryDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
...